// oasis/src/texas/types.gen.ts
export type IdentityProvider =
  | "azuread"
  | "tokenx"
  | "maskinporten"
  | "idporten";

/**
 * This data type holds the OAuth token that will be validated in the introspect endpoint.
 */
export type IntrospectRequest = {
  identity_provider: IdentityProvider;
  token: string;
};
// ...
export type IntrospectResponse = {
  /**
   * Indicates whether the token is valid. If this field is _false_,
   * the token is invalid and *must* be rejected.
   */
  active: boolean;
  /**
   * If the token is invalid, this field contains the reason.
   */
  error?: string | null;
  [key: string]: unknown | boolean | string | null | undefined;
};